Simple calculator v0.1 beta
A small, lightweight, feature-lacking calculator written in C++.

No GUI, only CLI.

Want new features/collab with me? Hit me up on Discord! donut#2008

(If you're mad that this is so simple, remember that this is still in alpha lol)

I'm still new to C++ programming, so making this was quite a task. I'd definitely like some helpers :)